requirements.txt
alabaster==0.7.12
asn1crypto==1.2.0
astroid==2.3.3
astropy==3.0.2
atomicwrites==1.3.0
attrdict==2.0.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.1
bitarray==1.1.0
bkcharts==0.2
bleach==3.1.0
bokeh==0.12.16
boto==2.49.0
bottle==0.12.13
Bottleneck==1.2.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.6.0.post1
seaborn==0.8.1
idna==2.8

oauthlib==2.1.0
matplotlib==3.0.3
pytz==2019.3
pybitflyer==0.1.9
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.0.0
Flask==1.1.1
line-bot-sdk==1.8.0
six==1.13.0

twitter==1.18.0

urllib3==1.24.2
yfinance==0.1.54
